pricing text inject with dynamic import
